[#31] Do not validate defaults for Django fields with choices
The reason here is that the source of our config values cannot
express complex types such as enum members, and that defaults
for Django fields with choices frequently include such types
(or callables returning such types). We are mainly about the
types of the values provided from outside, not of Django field
defaults (beyond null and blank), so we can leave the validation
of the default to Django.